rename sub.rs to email.rs
